+++ packages/admin-ui/src/index.ts
@@ -2,3 +2,8 @@ export { default as Breadcrumbs } from './breadcrumbs';
 export { default as NavigableRegion } from './navigable-region';
 export { default as Page } from './page';
 export { getAdminThemeColors } from './admin-theme-colors';
+export {
+	default as EntitiesSavedStates,
+	EntitiesSavedStatesExtensible,
+} from './entities-saved-states';
+export { useIsDirty as useEntitiesSavedStatesIsDirty } from './entities-saved-states/hooks/use-is-dirty';