+++ packages/boot/src/components/save-panel/index.tsx
@@ -3,8 +3,8 @@
  */
 import { useState } from '@wordpress/element';
 import { Modal } from '@wordpress/components';
-import { EntitiesSavedStates } from '@wordpress/editor';
 import { __ } from '@wordpress/i18n';
+import { EntitiesSavedStates } from '@wordpress/admin-ui';
 
 /**
  * Internal dependencies